fix: align ChipScript language and compiler with real CHIP-8 hardware
Breaking changes to the language:

  - Integer model: unsigned byte 0-255 (was claimed -128..127 signed,
    but codegen treated as bytes anyway - now honest & consistent)
  - Negative literals removed (fixes a-10 lexing bug)
  - Operator precedence tiers: or < and < comparison < additive <
    multiplicative (was flat equal-precedence)
  - Max variables: 14 (V0 scratch, VF flag)
  - Function calls allowed anywhere (nested, fn-to-fn) via caller-saves
    ABI using Fx55/Fx65 dump/restore with per-depth save areas
  - Recursion now enforced at compile time (cycle detection)
  - Max call depth: 16 (CHIP-8 stack limit) enforced
  - Div/mod by zero: compile-time error
  - Variable initializers: constant expressions only
  - Name uniqueness across vars/fns/sprites enforced
  - Function args typed int; builtin args type-checked
  - drawdigit(n) masked to 0-15 via AND 0x0F (defined behaviour)
  - ROM size limit 3584 bytes enforced
  - Even-address alignment for code (Cowgod spec)

Docs: SPECS.md, DOCUMENTATION.md, README.md fully synced.
Tests: all 192 passing across 6 test suites.

+++ src/ast.rs
@@ -61,7 +61,7 @@ pub struct ElseIf {
 // an expression produces a value
 #[derive(Debug, Clone)]
 pub enum Expr {
-    Int(i16),                        // 10
+    Int(u8),                         // 10 (unsigned byte 0–255)
     Bool(bool),                      // true / false
     Var(String),                     // a
     BinOp(Box<Expr>, Op, Box<Expr>), // a + b, a == b
